Added original code in zip file
The original class was called CInstantCamera*For*AppSrc. This is now obsolete, and replaced with CInstantCameraAppSrc.
The new class contains some things that were previously implemented outside of the class.
So, the original source is included in this zip file in case it is needed.
Note: Apologies in advance, as the source filenames changed with the name of the class!